Update export_template_manager.cpp
Points it to a redotengine URL.  Probably want to change depending on how our infrastructure is laid out following release.

I mainly wanted to raise awareness that this needs to be altered, preferably before a release (even if it points to a repo where there are no export templates available, just as long as it's not upstream servers).